float32 dtype

+++ main.py
@@ -86,7 +86,7 @@ class cancer_ml:
 
         X = self.data_pipe.image_only.X_train
 
-        X = np.reshape(X, (X.shape[0], X.shape[1]*X.shape[2])).astype('int8')
+        X = np.reshape(X, (X.shape[0], X.shape[1]*X.shape[2])).astype('float32')
 
         print(len(self.model.labels_))
